refactor: apply review suggestions - rename modules, extract triton kernels, move variants
